requirements.txt
python-dotenv
watchdog
pyperclip